test/requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.4.1
bcrypt==4.0.1
PyJWT==2.8.0